Move Nemotron cleanup to standardization
